+++ types.d.ts
@@ -119,6 +119,7 @@ export type SignalKey = string | number | object | symbol;
  */
 export interface ReadonlySignal<T> {
   get(): T;
+  subscribe(): T;
   readonly value: T;
   stop(): void;
   transform<U>(fn: (value: T) => U, key?: SignalKey): ReadonlySignal<U>;
@@ -160,6 +161,7 @@ export interface ReadonlySignal<T> {
 export class Signal<T> implements ReadonlySignal<T> {
   private constructor();
   get(): T;
+  subscribe(): T;
   readonly value: T;
   set(valueOrFn: T | ((current: T) => T)): void;
   stop(): void;